requirements.txt
Django==2.1.7
gunicorn==19.9.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
pytz==2018.9
djangorestframework>=3.9.1
django-rest-knox>=3.6.0
rest-social-auth>=2.0.2
markdown-3.0.1
django-filter-2.1.0
